Add files via upload
Your assignment is to create a bar-chart showing the total revenue and the percentage of revenue each store contributes to the total revenue. The input for the store revenue is assumed to be in dollars and will be read from a file (to start you can prompt the user to enter the values). I added an SFML example file that was given to me by my professor.